Move TestRetrierCtx to retrier_test.go
